requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
azure-core==1.30.0
azure-identity==1.15.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
Deprecated==1.2.14
fastapi==0.109.2
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
hyperframe==6.0.1
idna==3.6
importlib-metadata==6.11.0
iniconfig==2.0.0
microsoft-kiota-abstractions==1.2.0
microsoft-kiota-authentication-azure==1.0.0
microsoft-kiota-http==1.3.0
microsoft-kiota-serialization-json==1.0.1
microsoft-kiota-serialization-text==1.0.0
microsoftgraph-python==1.1.6
msal==1.26.0
msal-extensions==1.1.0
msgraph-core==1.0.0
msgraph-sdk==1.1.0
multidict==6.0.5
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
packaging==23.2
pendulum==3.0.0
pluggy==1.4.0
portalocker==2.8.2
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
PyMySQL==1.1.0
pytest==8.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pywin32==306
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.36.3
std-uritemplate==0.0.52
time-machine==2.13.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
uvicorn==0.27.0.post1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
